+++ Runtime/FrankenBit.Utilities/StateMachine.Transitions.cs
@@ -4,6 +4,7 @@
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
 
+using System;
 using System.Collections.Generic;
 
 using JetBrains.Annotations;
@@ -18,7 +19,7 @@ namespace FrankenBit.Utilities
         /// <summary>
         ///     Collection of transitions.
         /// </summary>
-        private sealed class Transitions : ITransitions
+        private sealed class Transitions
         {
             /// <summary>
             ///     Encapsulated list of transitions.
@@ -29,17 +30,16 @@ namespace FrankenBit.Utilities
             public override string ToString() =>
                 $"{_transitions.Count}";
 
-            /// <inheritdoc />
-            public IState FindTransition( IState state )
-            {
-                foreach ( ITransition transition in _transitions )
-                {
-                    IState targetState = transition.GetTargetState();
-                    if ( targetState != null ) return targetState;
-                }
-
-                return null;
-            }
+            /// <summary>
+            ///     Try to find an available transition.
+            /// </summary>
+            /// <returns>
+            ///     Target <seealso cref="IState" /> to which a transition is available
+            ///     or <see langword="null" /> if no transition is currently available.
+            /// </returns>
+            [CanBeNull]
+            public IState FindTransition() =>
+                FindTransition( t => t.HasCondition ) ?? FindTransition( t => !t.HasCondition );
 
             /// <summary>
             ///     Add a new <paramref name="transition" /> to the collection.
@@ -49,6 +49,30 @@ namespace FrankenBit.Utilities
             /// </param>
             internal void Add( [NotNull] ITransition transition ) =>
                 _transitions.Add( transition );
+
+            /// <summary>
+            ///     Find an available transition matching the supplied <paramref name="condition" />.
+            /// </summary>
+            /// <param name="condition">
+            ///     Condition that has to be met by the transition to be found.
+            /// </param>
+            /// <returns>
+            ///     Target state returned by a matching transition or <see langword="null" />
+            ///     if there was no available transition matching the supplied <paramref name="condition" />.
+            /// </returns>
+            [CanBeNull]
+            private IState FindTransition( [NotNull] Func<ITransition, bool> condition )
+            {
+                foreach ( ITransition transition in _transitions )
+                {
+                    if ( !condition( transition ) ) continue;
+
+                    IState targetState = transition.GetTargetState();
+                    if ( targetState != null ) return targetState;
+                }
+
+                return null;
+            }
         }
     }
 }